@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-const compileLib = require("../install/compileLib.js");
+const compileLib = require("../cjs/install/compileLib.js");
 const {argv} = process;
 compileLib.compileLib(argv);
