@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import * as compileLib from "../install/compileLib";
+import { compileLib } from "../src/install/compileLib";
 const {argv} = process;
-compileLib.compileLib(argv);
+compileLib(argv);
