@@ -1,8 +1,8 @@
-import { getResourcePath, drawRect, wait4key } from './utils';
-import fs from 'fs';
-import path from 'path';
-import { classNames } from './data/dnnCocoClassNames';
-import { extractResults, Prediction } from './dnn/ssdUtils';
+import fs from 'node:fs';
+import path from 'node:path';
+import { getResourcePath, drawRect, wait4key } from './utils.js';
+import { classNames } from './data/dnnCocoClassNames.js';
+import { extractResults, Prediction } from './dnn/ssdUtils.js';
 import {cv, Mat, Net, Vec3 } from '@u4/opencv4nodejs';
 
 if (!cv.xmodules || !cv.xmodules.dnn) {
@@ -59,7 +59,7 @@ const runDetectDishesExample = async (net: Net) => {
       fontSize: 0.8,
       color: classColors[className],
     })),
-    alpha
+    alpha,
   );
 
   Object.keys(classColors).forEach((className) => {
