@@ -1,7 +1,7 @@
-import { cv, getResourcePath, wait4key } from './utils';
-import fs from 'fs';
-import path from 'path';
+import fs from 'node:fs';
+import path from 'node:path';
 import { Mat } from '@u4/opencv4nodejs';
+import { cv, getResourcePath, wait4key } from './utils.js';
 
 async function main() {
   if (!cv.xmodules || !cv.xmodules.dnn) {
@@ -102,7 +102,7 @@ async function main() {
       img,
       { x: 0, y: 0 },
       predictions.map(p => ({ text: p, fontSize: 0.5, thickness: 1 })),
-      alpha
+      alpha,
     );
     cv.imshow('img', img);
     await wait4key();
