@@ -1,6 +1,7 @@
 import { Mat } from '@u4/opencv4nodejs';
-import { cv, getResourcePath } from '../utils';
-import { runVideoFaceDetection } from './commons';
+import { cv, getResourcePath } from '../utils.js';
+import { runVideoFaceDetection } from './commons.js';
+
 if (cv.version.minor === 4) {
   console.log('Warning: It seems like opencv 3.4 does not run the opencl version of detectMultiScale.');
 }
