@@ -1,5 +1,5 @@
 import { CV_32F, imread, Mat, Size } from '@u4/opencv4nodejs';
-import { getResourcePath } from './utils';
+import { getResourcePath } from './utils.js';
 
 // Ported from https://docs.opencv.org/2.4/doc/tutorials/gpu/gpu-basics-similarity/gpu-basics-similarity.html
 function getStructureSimilarity(i1: Mat, i2: Mat): number {
