@@ -1,5 +1,5 @@
 import { DescriptorMatch, FeatureDetector, Mat } from '@u4/opencv4nodejs';
-import { cv, getResourcePath, wait4key } from './utils';
+import { cv, getResourcePath, wait4key } from './utils.js';
 
 const matchFeaturesPass = (arg: { img1: Mat, img2: Mat, detector: FeatureDetector, matchFunc: (descs1: Mat, descs2: Mat) => DescriptorMatch[] }) => {
   const { img1, img2, detector, matchFunc } = arg;
@@ -17,7 +17,7 @@ const matchFeaturesPass = (arg: { img1: Mat, img2: Mat, detector: FeatureDetecto
   // only keep good matches
   const bestN = 40;
   const bestMatches = matches.sort(
-    (match1, match2) => match1.distance - match2.distance
+    (match1, match2) => match1.distance - match2.distance,
   ).slice(0, bestN);
 
   return cv.drawMatches(
@@ -25,7 +25,7 @@ const matchFeaturesPass = (arg: { img1: Mat, img2: Mat, detector: FeatureDetecto
     img2,
     keyPoints1,
     keyPoints2,
-    bestMatches
+    bestMatches,
   );
 };
 
