@@ -1,11 +1,9 @@
-
-import { cv } from "./utils";
-
+import { cv } from './utils.js';
 
 try {
   console.log('');
   console.log('1 Dims');
-  const mat1 = new cv.Mat([4, 5, 6] as any, cv.CV_8UC3);
+  const mat1 = new cv.Mat([4, 5, 6], cv.CV_8UC3);
   //const mat1 = new cv.Mat([[
   //    [1, 2, 3.3],
   //    [4, 5, 6],
@@ -19,7 +17,7 @@ try {
 try {
   console.log('');
   console.log('2 Dims');
-  const mat2 = new cv.Mat([[4, 5, 6]] as any, cv.CV_32FC3);
+  const mat2 = new cv.Mat([[4, 5, 6]], cv.CV_32FC3);
   //const mat1 = new cv.Mat([[
   //    [1, 2, 3.3],
   //    [4, 5, 6],
@@ -34,7 +32,7 @@ try {
 try {
   console.log('');
   console.log('3 Dims');
-  const mat3 = new cv.Mat([[[4, 5, 6]]] as any, cv.CV_32FC3);
+  const mat3 = new cv.Mat([[[4, 5, 6]]], cv.CV_32FC3);
   //const mat1 = new cv.Mat([[
   //    [1, 2, 3.3],
   //    [4, 5, 6],
@@ -42,14 +40,13 @@ try {
   //
   console.log(mat3.getDataAsArray());
 } catch (e) {
-
+  // empty
 }
-
 
 try {
   console.log('');
   console.log('4 Dims');
-  const mat4 = new cv.Mat([[[[4, 5, 6]]]] as any, cv.CV_32FC3);
+  const mat4 = new cv.Mat([[[[4, 5, 6]]]], cv.CV_32FC3);
   //const mat1 = new cv.Mat([[
   //    [1, 2, 3.3],
   //    [4, 5, 6],
@@ -57,5 +54,5 @@ try {
   //
   console.log(mat4.getDataAsArray());
 } catch (e) {
-
+  // empty
 }
