@@ -9,14 +9,15 @@ const isAsyncFn = (fn: (...args: any[]) => any) => {
   }
 };
 
-const promisify = (fn: () => any) => function (...params: any[]) {
+const promisify = (fn: (...args: any[]) => any) => function toPromise(this: any, ...params: any[]) {
   if (isFn(params[params.length - 1])) {
-    return fn.apply(this, params);
+    return fn.apply(this, params as Parameters<typeof fn>);
   }
 
   return new Promise((resolve, reject) => {
     const args = Array.prototype.slice.call(params);
-    args.push(function(err: Error, res: any) {
+    // add a param to convert callback to promise
+    args.push(function pomisifyed(err: Error, res: any) {
       if (err) {
         return reject(err);
       }
@@ -27,7 +28,7 @@ const promisify = (fn: () => any) => function (...params: any[]) {
   });
 };
 
-export default <T>(cv: T): T => {
+export default <T extends { [key: string]: any }>(cv: T): T => {
   const fns = Object.keys(cv).filter(k => isFn(cv[k])).map(k => cv[k]);
   const asyncFuncs = fns.filter(isAsyncFn);
   const clazzes = fns.filter(fn => fn.prototype && !!Object.keys(fn.prototype).length);
@@ -38,7 +39,8 @@ export default <T>(cv: T): T => {
   });
 
   asyncFuncs.forEach((fn) => {
-    cv[fn.prototype.constructor.name] = promisify(fn);
+    const {name} = fn.prototype.constructor;
+    (cv as any)[name] = promisify(fn);
   });
 
   return cv;
