@@ -2,8 +2,9 @@ import fs from 'fs';
 import path from 'path';
 import { expect } from 'chai';
 import { Mat } from '@u4/opencv4nodejs';
-import { getTestContext } from '../model';
-import { assertDataDeepEquals, assertMetaData } from '../../utils/matTestUtils';
+import { getDirName } from '@u4/opencv4nodejs';
+import { getTestContext } from '../model.js';
+import { assertDataDeepEquals, assertMetaData } from '../../utils/matTestUtils.js';
 import {
   clearTmpData,
   fileExists,
@@ -11,9 +12,9 @@ import {
   getTmpDataFilePath,
   _asyncFuncShouldRequireArgs,
 }
-  from '../../utils/testUtils';
-import { generateAPITests } from '../../utils/generateAPITests';
-import toTest from '../toTest';
+  from '../../utils/testUtils.js';
+import { generateAPITests } from '../../utils/generateAPITests.js';
+import toTest from '../toTest.js';
 
 if (toTest.io) {
   const {
@@ -36,8 +37,8 @@ if (toTest.io) {
   before(() => {
     lenna = cv.imread(getLennaPngPath());
     got = cv.imread(getGotJpgPath());
-    const lennaBase64File = fs.readFileSync(path.join(__dirname, 'data/lennaBase64.json'), { encoding: 'utf8', flag: 'r' });
-    const gotBase64File = fs.readFileSync(path.join(__dirname, 'data/gotBase64.json'), { encoding: 'utf8', flag: 'r' });
+    const lennaBase64File = fs.readFileSync(path.join(getDirName(), '../../test/tests/io/data/lennaBase64.json'), { encoding: 'utf8', flag: 'r' });
+    const gotBase64File = fs.readFileSync(path.join(getDirName(), '../../test/tests/io/data/gotBase64.json'), { encoding: 'utf8', flag: 'r' });
     lennaBase64Buf = Buffer.from(JSON.parse(lennaBase64File).data, 'base64');
     gotBase64Buf = Buffer.from(JSON.parse(gotBase64File).data, 'base64');
     // imageData = fs.readFileSync(getTestImagePath(true));
