@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable no-unused-vars */
+ 
 import { assert, expect } from 'chai';
 import fs from 'fs';
 import {
@@ -8,7 +8,6 @@ import {
   Vec3,
   Vec4,
 } from '@u4/opencv4nodejs';
-import type Chai from 'chai';
 
 export const assertError = (func: () => any, msg: string): void => {
   let errMsg = '';
