@@ -86,7 +86,7 @@ export const DNN_BACKEND_OPENCV: number;
 export const DNN_BACKEND_INFERENCE_ENGINE: number;
 export const DNN_BACKEND_HALIDE: number;
 export const DNN_BACKEND_CUDA: number;
-
+export const DNN_BACKEND_WEBNN: number;
 export const DNN_TARGET_CPU: number;
 export const DNN_TARGET_OPENCL: number;
 export const DNN_TARGET_OPENCL_FP16: number;
@@ -634,6 +634,7 @@ export const INTER_CUBIC: number;
 export const INTER_LANCZOS4: number;
 export const INTER_LINEAR: number;
 export const INTER_MAX: number;
+export const INTER_LINEAR_EXACT: number;
 export const INTER_NEAREST: number;
 export const KMEANS_PP_CENTERS: number;
 export const KMEANS_RANDOM_CENTERS: number;
@@ -657,6 +658,7 @@ export const MORPH_RECT: number;
 export const MORPH_TOPHAT: number;
 export const NORM_HAMMING: number;
 export const NORM_HAMMING2: number;
+export const NORM_TYPE_MASK: number;
 export const NORM_INF: number;
 export const NORM_L1: number;
 export const NORM_L2: number;
@@ -670,6 +672,9 @@ export const REGULAR: number;
 export const RETR_CCOMP: number;
 export const RETR_EXTERNAL: number;
 export const RETR_FLOODFILL: number;
+export const CONTOURS_MATCH_I1: number;
+export const CONTOURS_MATCH_I2: number;
+export const CONTOURS_MATCH_I3: number;
 export const RETR_LIST: number;
 export const RETR_TREE: number;
 export const RHO: number;
